+++ transformer-runner/src/types.ts
@@ -22,6 +22,12 @@ interface ArtifactData extends ContractData {
 		slugSuffix?: string; // used to allow transformers customization of generated slugs. needed when creating multiple instances of same type
 		encryptedSecrets?: any;
 		backflow?: ArtifactContract[];
+		repoData?: {
+			org: string;
+			repo: string;
+			branch: string;
+			commit: string;
+		};
 	};
 }
 export interface ArtifactContract extends Contract<ArtifactData> {}